+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For more information about changelogs, check
 * [breaking change] - CamelCaser module updated to apply #camel_case to lists
 * [breaking change] - Dasher module updated to apply #dasherize to lists
 * [breaking change] - LowerCaser module updated to apply #lowercase_first to lists
+* [breaking change] - SnakeCaser module updated to apply #snake_case to lists
 
 ## 1.0.0 - 2/17/18 Sat